[Hack][DASHTree] Stop full update when segment wasn't found
Sometimes server sends an mpd from the past and we are not able to find
a corresponding segment for the update. Currently we decrement a start number
by it causes a playback to rewind for a few seconds.

Implement a hack which drops that invalid mpd and hope that server will
send a proper one in the next update.